fix: adjusting home page to be the main page of this site
