+++ src/app/layout.js
@@ -5,7 +5,7 @@ import "./globals.css"
 
 export default function RootLayout({ children, session }) {
   return (
-    <html lang="en">
+    <html lang="pt-BR">
       <head>
         <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
       </head>
